# src/schema/openapi.specification.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication service",
    "version": "1.0.0"
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Get a list of users",
        "description": "Retrieve a list of user entities.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Resource to create a user",
        "description": "Create a user",
        "requestBody": {
          "description": "User to add to the system",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Resource to retrieve a single user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User identification string",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Resource to delete a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User identification string",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response"
          }
        }
      },
      "put": {
        "summary": "Resource to update user",
        "requestBody": {
          "description": "User to update io the system",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User identification string",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserDTO": {
        "type": "object",
        "required": ["email", "firstName", "lastName"],
        "properties": {
          "email": {
            "type": "string",
            "default": "johndoe@gmail.com"
          },
          "firstName": {
            "type": "string",
            "default": "john"
          },
          "lastName": {
            "type": "string",
            "default": "Doe"
          },
          "isFirstTimeSignin": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "UserInput": {
        "type": "object",
        "required": ["email", "firstName", "lastName", "password"],
        "properties": {
          "email": {
            "type": "string",
            "default": "johndoe@gmail.com"
          },
          "password": {
            "type": "string",
            "default": "securePassword123"
          },
          "firstName": {
            "type": "string",
            "default": "john"
          },
          "lastName": {
            "type": "string",
            "default": "Doe"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "default": 123
          },
          "message": {
            "type": "string",
            "default": "Something went wrong"
          },
          "error": {
            "type": "object"
          }
        }
      }
    }
  },
  "tags": []
}